fix: dev: Fix Taiga Bug #487 NTP: Connection not established with NTP server using physical switch
NTPD does not synchronize time if the local time is skewed by more than
1000 secs.
In case of switches, where you have a default configuration and you're
setting it up for the first time - there can be instances where this
should not be the expected behavior. For which we'll have to
specifically tell NTPD to support such scenarios. This fix provides a
mechanism  to tell NTPD to not misbehave in those situations.

Change-Id: I6435a5d7394f2bf2a757286def08991e15962612
Signed-off-by: srinsriv <srinivasan.srivatsan@hpe.com>