fix: dev: Fix Bug #429 Python ntpd cpu spike issue
Root Cause:
Init portion of the python ntpd handles registering to OVSDB and
fetching idl information. This is the portion which was hogging
CPU. Previously we had the ops_ntpd.py which called ops_ntpd_sync_to_ovsdb.py
very frequently for the updates. Each call meant that we reinit
and register to OVSDB. Because of which we see high cpu spike every
few seconds.

Solution:
Decouple the init portion of ops_ntpd_sync_to_ovsdb.py and have it
initialize only once. For the subsequent status updates, only create
and submit a transaction.

Change-Id: I950cbedb0499414693876dfa93a227649fd879e0
Signed-off-by: srinsriv <srinivasan.srivatsan@hpe.com>